Update Genre and Date
